Extract posting component into some individual components
