+++ src/components/posting/love-shape/LoveShape.scss
@@ -1,0 +1,6 @@
+.like-and-heart-icon {
+  width: 24px;
+  height: 24px;
+  margin-right: 5px;
+  cursor: pointer;
+}
