+++ src/components/posting/post-bottom-section/PostBottomSection.scss
@@ -1,0 +1,17 @@
+.post-bottom {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 30px;
+  box-sizing: border-box;
+
+  .post-bottom-left {
+    display: flex;
+    align-items: center;
+  }
+
+  .post-bottom-right {
+    font-size: 12px;
+    font-weight: bold;
+  }
+}
