+++ src/components/posting/post-container/PostContainer.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import "./PostContainer.scss"
+
+export default function PostContainer({ children }) {
+  return <div className="post">{children}</div>;
+}
