+++ src/components/posting/post-container/PostContainer.scss
@@ -1,0 +1,8 @@
+.post {
+  width: 100%;
+  border-radius: 10px;
+  -webkit-box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+  box-shadow: 0px 0px 16px -8px rgba(0, 0, 0, 0.68);
+  margin: 10px 0;
+  background-color: white;
+}
