+++ src/components/posting/post-like-counter/PostLikeCounter.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { displayUserWhoLikesThisPost } from "../../../utils/postLikes.js";
+
+import "./PostLikeCounter.scss";
+
+export default function PostLikeCounter({
+  totalLikeCalculated,
+  postedDataLikesMapped,
+  currentUserIdSlice,
+}) {
+  return (
+    <span className="post-like-counter">
+      {totalLikeCalculated},{" "}
+      {displayUserWhoLikesThisPost(postedDataLikesMapped, currentUserIdSlice)}
+    </span>
+  );
+}
