+++ src/components/posting/post-like-counter/PostLikeCounter.scss
@@ -1,0 +1,7 @@
+.post-like-counter {
+  font-size: 9px;
+
+  @media only screen and (min-width: 1024px) {
+    font-size: 10px;
+  }
+}
