+++ src/components/posting/post-middle-section/PostMiddleSection.scss
@@ -1,0 +1,15 @@
+.post-middle {
+  margin: 14px 0;
+
+  .post-middle-text {
+    font-size: 12px;
+  }
+
+  .post-img {
+    width: 100%;
+    max-height: 500px;
+    object-fit: contain;
+    margin-top: 20px;
+    border-radius: 4px;
+  }
+}
