+++ src/components/posting/post-top-section/PostTopSection.scss
@@ -1,0 +1,82 @@
+.post-top {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+
+  .post-top-left {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+
+    .post-profile-img-link-wrapper {
+      width: 32px;
+      height: 32px;
+      border-radius: 50%;
+      cursor: pointer;
+
+      .post-profile-img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        border-radius: 50%;
+      }
+    }
+
+    .post-top-left-username-and-date-wrapper {
+      padding-left: 10px;
+      display: flex;
+      flex-direction: column;
+      gap: 4px;
+
+      .post-username {
+        font-size: 13px;
+        font-weight: 500;
+        cursor: pointer;
+      }
+
+      .post-date-and-status-wrapper {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+
+        .post-date {
+          font-size: 10px;
+          margin-right: 6px;
+        }
+      }
+    }
+  }
+
+  .post-top-right {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    .post-status {
+      font-size: 8px;
+      padding: 4px 6px;
+      border-radius: 6px;
+
+      @media only screen and (min-width: 1024px) {
+        font-size: 10px;
+      }
+
+      &.public {
+        background-color: rgb(141, 174, 243);
+        color: white;
+      }
+
+      &.private {
+        background-color: rgb(241, 93, 212);
+        color: white;
+      }
+
+      &.followers-only {
+        background-color: rgb(32, 202, 217);
+        color: white;
+      }
+    }
+  }
+}
