+++ src/components/posting/post-wrapper/PostWrapper.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import "./PostWrapper.scss";
+
+export default function PostWrapper({ children }) {
+  return <div className="post-wrapper">{children}</div>;
+}
