+++ src/components/posting/post-wrapper/PostWrapper.scss
@@ -1,0 +1,5 @@
+.post-wrapper {
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+}
